# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: NoteHub API
  version: "0.1.0"
  description: |
    NoteHub REST API および WebSocket 仕様。

    ## 認証
    - Access Token: `Authorization: Bearer <token>` ヘッダー（15分有効）
    - Refresh Token: HttpOnly Cookie `notehub_refresh_token`（`/api/auth` パス）
    - CSRF: Cookie `notehub_csrf_token` と Header `X-CSRF-Token` の Double Submit（`POST /api/auth/refresh`, `POST /api/auth/logout`）

    ## レスポンス形式
    成功時は `{ "data": ... }`、失敗時は `{ "error": { "code", "message", "requestId?" } }`。
servers:
  - url: http://localhost:8080
    description: Local development
  - url: http://localhost:8082
    description: Local integration test
...
paths:
...
  /api/workspaces:
    get:
      tags: [Workspace]
      summary: ワークスペース一覧
      operationId: listWorkspaces
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/WorkspaceListItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Workspace]
      summary: ワークスペース作成
      operationId: createWorkspace
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/CreateWorkspaceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/workspaces/{workspaceId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      tags: [Workspace]
      summary: ワークスペース詳細
      operationId: getWorkspace
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/GetWorkspaceResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Workspace]
      summary: ワークスペース更新（ホストのみ）
      operationId: updateWorkspace
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UpdateWorkspaceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    delete:
      tags: [Workspace]
      summary: ワークスペース削除（ホストのみ）
      operationId: deleteWorkspace
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteWorkspaceRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/DeleteWorkspaceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access Token（15分有効、メモリ保持）
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF Cookie `notehub_csrf_token` と同値
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Too Many Requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    WorkspaceForbidden:
      description: Workspace access or permission denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ApiResponse:
      type: object
      required: [data]
      properties:
        data: {}
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: 入力値が不正です
            requestId:
              type: string
              format: uuid
...
    CreateWorkspaceRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100
    CreateWorkspaceResponse:
      type: object
      required: [id, name, hostId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        hostId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    WorkspaceListItem:
      type: object
      required: [id, name, hostId, hostName, role, isAvailable, unavailableReason, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        hostId:
          type: string
          format: uuid
        hostName:
          type: string
        role:
          type: string
          enum: [host, member]
        isAvailable:
          type: boolean
        unavailableReason:
          type: string
        updatedAt:
          type: string
          format: date-time
    WorkspaceHost:
      type: object
      required: [id, name, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [active, suspended, deleted]
    GetWorkspaceResponse:
      type: object
      required: [id, name, host, role, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        host:
          $ref: "#/components/schemas/WorkspaceHost"
        role:
          type: string
          enum: [host, member]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateWorkspaceRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100
    UpdateWorkspaceResponse:
      type: object
      required: [id, name, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        updatedAt:
          type: string
          format: date-time
    DeleteWorkspaceRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          maxLength: 500
    DeleteWorkspaceResponse:
      type: object
      required: [workspaceId, deletedAt]
      properties:
        workspaceId:
          type: string
          format: uuid
        deletedAt:
          type: string
          format: date-time
...